requirements.txt
Flask==2.3.2
tensorflow==2.17.0
numpy==1.24.4
Pillow==10.0.0
